requirements.txt
pandas
numpy
scikit-learn
nltk
spacy
PyPDF2
python-docx
